fix(columnar): widen numValues and ordinal to long
Aggregate numValues in a ColumNAR numeric column is not bounded by
maxDoc: multi-valued docs pack multiple values per binary payload, so
the total across all documents can exceed Integer.MAX_VALUE.

numValues in NumericColumnMetadata widened to long; wire encoding
switched from writeVInt/readVInt to writeVLong/readVLong. The ordinal
accumulator in NumericColumnWriter widened to long; a wrapped int
would corrupt firstOrdinal() for every doc after 2^31 values.
numBlocks() and block-index types follow. Per-doc valueCount() stays int.

Adds testLargeNumValuesMetadataRoundTrips: draws numValues from
[Integer.MAX_VALUE+1, 2*Integer.MAX_VALUE], round-trips the metadata
through writeTo/readFrom, and asserts the value survives intact.
Without the fix: expected:<2179787566> but was:<-2115179730>.

+++ libs/columnar/src/main/java/org/elasticsearch/columnar/numeric/ColumnarNumericBinaryDocValues.java
@@ -67,7 +67,7 @@ public final class ColumnarNumericBinaryDocValues extends BinaryDocValues {
     @Override
     public BytesRef binaryValue() throws IOException {
         final int rank = iterator.index();
-        final int first = reader.firstOrdinal(rank);
+        final long first = reader.firstOrdinal(rank);
         final int count = reader.valueCount(rank);
         if (values.length < count) {
             values = new long[ArrayUtil.oversize(count, Long.BYTES)];
@@ -114,7 +114,7 @@ public final class ColumnarNumericBinaryDocValues extends BinaryDocValues {
      */
     public NumericColumnValues directValues() {
         return new NumericColumnValues() {
-            private int first;
+            private long first;
             private int count;
             private int upto;
